refactored ShowMe into its own Component
